Revert "SI-1786 incorporate defined bounds in inference"
Have to revert because the stricter bounds that it inferred break e.g., slick.
(Backstop for that added as pos/t1786-counter.scala, as minimized by Jason)
Worse, the fix was compilation order-dependent.

There's a less invasive fix (SI-6169) that could be generalized
in `sharpenQuantifierBounds` (used in `skolemizeExistential`),
but I'd rather not mess with existentials at this point.

This reverts commit e28c3edda4dd405ed382227d2a688b799bf33c72.

Conflicts:
	src/compiler/scala/tools/nsc/typechecker/Typers.scala
	test/files/pos/t1786.scala