SI-7932 Exclude PolyTypes from Java generic signatures
In the enclosed test case, we were emitting just the result type of
`[a, b]Float` in the Java generic signature.

This resulted in a `GenericSignatureFormatError`.

This commit changes `argSig` to project such type functions to `*`
instead.

The test case shows that we still emit the class when we use
its type constructor directly as the type argument.