SI-8325 Accept infix star type outside patterns
This is a follow-up to SI-5702 which enabled use of
`*` in infix notation in patterns.

Most of the work is in distinguishing infix from a
sequence pattern.

Also, do not take backticked star as the repeated
parameter marker in postfix position. That is,
`Int``*``` is not `Int*` -- I hope double-tick
renders as tick. There is not a special use case
except that backticks mean "I am an identifier, as
is, and not a keyword."