SI-8266 Amend advice for deprecated octal 042
Improve the advice for `f"\042"` to read:

```
use ${'"'} or a triple-quoted literal """with embedded " or \u0022""" instead.
```

as per the discussion on SI-6476.

Knuth says that Charles XII came close to introducing octal arithmetic to Sweden,
and Wikipedia doesn't deny it.

I imagine an alternative history in which octal literals are deprecated in Scala
but required by legislation in Akka. #octal-fan-fiction