SI-4492 More informative error when class not found on classpath
Position the error based on Select tree that failed to type check,
presumably due to an underlying MissingRequirementError, which has no position.

There are lots of other ways we could rewrap a MRE and supplement position info,
but that remains TODO. Jason's review comment is recorded in the code.

Also try to detect the case of a missing module and provide some advice,
as well as linking to the forthcoming 2.11 guide at
http://docs.scala-lang.org/overviews/core/scala-2.11.html.