SI-7992 Fix super-accessor generation after a local class
The transformer in the superaccessors phase uses the var
`validCurrentOwner` to track whether we're in a part of the
code that won't end up in the host method, and as such, will
need to access super-method via a super-accessor.

But, this bit of bookkeeping was not correctly reset after traversing
out of a local class. A `VerifyError` ensued.

This commit changes `atOwner` to save and restore that flag, rather
than leaving it set to `true`.

I've also added a test variation using a by-name argument.