Fix #6: automatic section numbering.
Add chapter number to YAML, which is replace in numbering.css.
Use CSS counters for chapters, sections, subsections | examples.
Examples are detected by looking at the H3 element's id,
which redcarpet derives from the heading's title.
It must start with "Example:", and the whole title is suppressed
by a little javascript, so we can make Examples look like in the pdf.

For example, `### Example:` becomes `Example 3.2.10`,
if it's the 10th example in Section 3.2.